src/project/requirements.txt
Django>=4.0
psycopg2-binary>=2.9
requests
python-docx
ollama
